requirements.txt
lxml==5.1.0
psycopg[binary]==3.1.18
pydantic==2.6.4
python-dotenv==1.0.1
pandas>=2.0.0
openpyxl>=3.1.2
